refactor: fix todo
